Fix bad interaction between switching static/dynamic trees and migration.
When we detect that something should move between the static and
dynamic trees, we save a record and process the move on the next
tick. However, it is possible the object will migrate in the meantime,
leaving a record of the switch but no object to operate on. Removing
the record when the object migrates would be ideal, but is a difficult
change to make because of synchronization. Instead, expose a function
in the location service cache to check whether an object still exists
or not and only process the move if the object is still available.

This also adjusts CBRLocationServiceCache to track existence as a
reference count to handle migrations. Because events can come in from
different servers in the "wrong" order (i.e. get replica from server
B where object O has moved to, then get the replica removal from
server A where it moved from), we need to use counts to track whether
we're aware of *any* valid copies of an object since they all share
one record in the location service cache.